Add PostgreSQL database connection in db.go
A new file, db.go, was created to establish a connection with a PostgreSQL database. A ConnectDB function was written which attempts to open a connection using a connection string. If the environment variable AutoMigrateDb is set to "true", the database is automatically migrated using Go's gorm.Open functionality. This change is done to improve the scalability of the database operations and handle concurrency better in a production environment.